+++ playground/src/Main.fs
@@ -4,6 +4,7 @@ open Feliz
 open Browser.Dom
 
 let private root = ReactDOM.createRoot(document.getElementById "root")
+
 root.render(
     React.StrictMode [
         App.App()